// features/payment/payment.dto.ts
import { $Enums } from '@/lib/generated/prisma';
import PaymentProvider = $Enums.PaymentProvider;
import { z } from 'zod';
import PaymentStatus = $Enums.PaymentStatus;
import Currency = $Enums.Currency;

export const CreatePaymentSchema = z.object({
  provider: z.enum(PaymentProvider),
  method: z.string(),
  amount: z.number().positive(),
  status: z.enum(PaymentStatus).default('PENDING'),
  currency: z.enum(Currency),
  externalId: z.string().optional(),
  rawPayload: z.record(z.string(), z.any()).optional(),
});

export type CreatePaymentInput = z.infer<typeof CreatePaymentSchema>;